WignerDMatrices.jl

WignerDMatrices.WignerDMatrix — Type

WignerDMatrix(j, α::Real, β::Real, γ::Real)

Construct the Wigner D-matrix in the z-y-z convention for the angular momentum j and the Euler angles (α, β, γ). The angular momentum j needs to be an Integer or a half-integer. The resulting matrix has a size of (2j+1, 2j+1) and axes (-j:j, -j:j).

Examples

julia> D = WignerDMatrix(0.5, 0, 0, 0)
2×2 WignerDMatrix{Complex{Float64}} for j = 1/2, alpha = 0, beta = 0 and gamma = 0 with indices -1/2:1/2×-1/2:1/2:
 1.0+0.0im  0.0-0.0im
 0.0+0.0im  1.0+0.0im

julia> D[-0.5, 0.5]
-0.0 + 0.0im

WignerDMatrices.WignerDMatrix — Method

WignerDMatrix{T}(j, α::Real, β::Real, γ::Real) where {T<:Complex}

Construct the Wigner D-matrix in the z-y-z convention for the angular momentum j and the Euler angles (α, β, γ), where the matrix elements are of the type T.

Examples

julia> WignerDMatrix{Complex{BigFloat}}(0.5, 0, 0, 0)
2×2 WignerDMatrix{Complex{BigFloat}} for j = 1/2, alpha = 0, beta = 0 and gamma = 0 with indices -1/2:1/2×-1/2:1/2:
 1.0+0.0im  0.0-0.0im
 0.0+0.0im  1.0+0.0im

WignerDMatrices.WignerdMatrix — Type

WignerdMatrix(j, beta::Real)

Compute the Wigner d-matrix in the z-y-z convention for the angular momentum j and the angle beta. The resulting matrix has a size of (2j+1, 2j+1) and axes (-j:j, -j:j). The angular momentum j needs to be either an Integer or a half-integer. The matrix elements are real in this convention.

Examples

julia> d = WignerdMatrix(0.5, π)
2×2 WignerdMatrix{Float64} for j = 1/2 and beta = π with indices -1/2:1/2×-1/2:1/2:
  0.0  1.0
 -1.0  0.0

julia> d[0.5, 0.5]
0.0

WignerDMatrices.WignerdMatrix — Method

WignerdMatrix{T}(j, β::Real) where {T<:Real}

Construct the Wigner d-matrix in the z-y-z convention for the angular momentum j and the angle β, where the elements of the matrix will be of the type T.

The matrix elements are expanded in the basis of the eigenfunctions of the angular momentum operator Jy. Specifying the type of the matrix elements does not increase the accuracy of the numerical evaluation of eigenvectors, as this is set by the LAPACK implementation used. An increased precision, however, might help in avoiding round-off errors in the sum.

Examples

julia> d = WignerdMatrix{BigFloat}(1, 0.4)
3×3 WignerdMatrix{BigFloat} for j = 1 and beta = 0.4 with indices -1:1×-1:1:
  0.96053     0.27536   0.0394695
 -0.27536     0.921061  0.27536
  0.0394695  -0.27536   0.96053

julia> d[1,1]
0.9605304970014426312106516543295635109444272950721034465148443857827477926800066

julia> WignerdMatrix{Float64}(1, 0.4)[1,1] # compare with Float64
0.9605304970014427

WignerDMatrices.wignerDmatrixelement — Function

wignerDmatrixelement(j::Real, m, n, (α, β, γ)::NTuple{3,Real}, [v = WignerDMatrices.eigenvecsJy!(j)])

Compute the element $D^j_{m,n}(α, β, γ)$ of the Wigner D-matrix in the z-y-z convention. Optionally the matrix v that contains the transposed eigenvectors of the angular momentum operator Jy corresponding to j may be provided.

The angular momentum j must be either an Integer or a half-integer.

Examples

julia> Dmn = WignerDMatrices.wignerDmatrixelement(1/2, 1/2, 1/2, (0, WignerDMatrices.TwoPi, 0))
-1.0 + 0.0im

julia> D = WignerDMatrix(1/2, 0, WignerDMatrices.TwoPi, 0)
2×2 WignerDMatrix{Complex{Float64}} for j = 1/2, alpha = 0, beta = 2π and gamma = 0 with indices -1/2:1/2×-1/2:1/2:
 -1.0-0.0im       ⋅    
      ⋅      -1.0-0.0im

julia> D[1/2,1/2] == Dmn
true

WignerDMatrices.wignerdmatrixelement — Function

wignerdmatrixelement(j::Real, m, n, β::Real, [v = WignerDMatrices.eigenvecsJy!(j)])

Compute the element $d^j_{m,n}(β)$ of the Wigner d-matrix in the z-y-z convention. Optionally the matrix v that contains the transposed eigenvectors of the angular momentum operator Jy corresponding to j may be provided.

The angular momentum j must be either an Integer or a half-integer.

Examples

julia> dmn = WignerDMatrices.wignerdmatrixelement(1/2, 1/2, 1/2, WignerDMatrices.TwoPi)
-1.0

julia> d = WignerdMatrix(1/2, WignerDMatrices.TwoPi)
2×2 WignerdMatrix{Float64} for j = 1/2 and beta = 2π with indices -1/2:1/2×-1/2:1/2:
 -1.0    ⋅ 
   ⋅   -1.0

julia> d[1/2,1/2] == dmn
true